Moving some functions and adding final functionality for active submission table endpoint
